+++ horovod/common/ops/gloo_operations.cc
@@ -264,8 +264,7 @@ Status GlooAllgather::Execute(std::vector<TensorTableEntry>& entries,
 
 
   timeline.ActivityStartAll(entries, ALLOCATE_OUTPUT);
-  Status status =
-      AllocateOutput(entries, response, entry_component_sizes, recvcounts);
+  Status status = AllocateOutput(entries, response, entry_component_sizes);
   if (!status.ok()) {
     /* Cleanup */
     for (size_t ec = 0; ec < entries.size(); ++ec) {
@@ -280,9 +279,10 @@ Status GlooAllgather::Execute(std::vector<TensorTableEntry>& entries,
   }
   timeline.ActivityEndAll(entries);
 
+  SetRecvcounts(entry_component_sizes, entries.size(), global_size, recvcounts);
   SetDisplacements(recvcounts, displcmnts, global_size);
-  SetEntryComponentOffsets(entries, entry_component_sizes, recvcounts,
-                           entry_component_offsets);
+  SetEntryComponentOffsets(entry_component_sizes, recvcounts, entries.size(),
+                           global_size, entry_component_offsets);
 
   std::unique_ptr<IGlooAlgorithms> gloo_algos(
       GetAlgorithmsForType(first_entry.tensor->dtype(), &gloo_context));